Appender: support primitive lists (1.4)
This is a backport of the PR #398 to `v1.4-andium` stable branch.

This change allows to append to primitive `LIST` columns using the same
methods that were used to append to primitive `ARRAY` columns.

Appending to nested lists or lists of arrays is not yet supported.

Testing: existing tests restructured and extended to cover primitive
`LIST`s.
