+++ src/test/java/org/duckdb/TestAppender.java
@@ -801,771 +801,6 @@ public class TestAppender {
         }
     }
 
-    public static void test_appender_array_basic() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 INTEGER[3])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                appender.beginRow().append(41).append(new int[] {41, 42, 43}).endRow();
-                appender.beginRow()
-                    .append(42)
-                    .append(new int[] {44, 45, 46}, new boolean[] {false, true, false})
-                    .endRow();
-                appender.beginRow().append(43).appendNull().endRow();
-                assertEquals(appender.flush(), 3L);
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT unnest(col2) FROM tab1 WHERE col1 = 41")) {
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 41);
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 42);
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 43);
-                assertFalse(rs.next());
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT unnest(col2) FROM tab1 WHERE col1 = 42")) {
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 44);
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 0);
-                assertTrue(rs.wasNull());
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 46);
-                assertFalse(rs.next());
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col2 FROM tab1 WHERE col1 = 43")) {
-                assertTrue(rs.next());
-                assertNull(rs.getObject(1));
-                assertTrue(rs.wasNull());
-                assertFalse(rs.next());
-            }
-        }
-    }
-
-    public static void test_appender_array_bool() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;         // auto flush
-            int tail = 16;               // flushed on close
-            int arrayLen = (1 << 6) + 6; // increase this for stress tests
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 BOOLEAN[" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    boolean[] arr = new boolean[arrayLen];
-                    for (byte j = 0; j < arrayLen; j++) {
-                        arr[j] = true;
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs =
-                     stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getBoolean(2), true);
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery("SELECT unnest(col2) FROM tab1 WHERE col1 = " + row)) {
-                for (byte j = 0; j < arrayLen; j++) {
-                    assertTrue(rs2.next());
-                    assertEquals(rs2.getBoolean(1), true);
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_nested_array_bool() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;         // auto flush
-            int tail = 17;               // flushed on close
-            int arrayLen = (1 << 6) + 5; // increase this for stress tests
-            int childLen = (1 << 6) + 7;
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 BOOLEAN[" + childLen + "][" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    boolean[][] arr = new boolean[arrayLen][childLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        for (int k = 0; k < childLen; k++) {
-                            arr[j][k] = 0 == (i + j + k) % 2;
-                        }
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "][" + (childLen - 1) +
-                                                  "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getBoolean(2), 0 == (i + arrayLen + childLen - 4) % 2);
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery(
-                     "SELECT unnest(ucol2) FROM (SELECT unnest(col2) as ucol2 FROM tab1 WHERE col1 = " + row + ")")) {
-                for (int j = 0; j < arrayLen; j++) {
-                    for (int k = 0; k < childLen; k++) {
-                        assertTrue(rs2.next());
-                        assertEquals(rs2.getBoolean(1), 0 == (row + j + k) % 2);
-                    }
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_array_tinyint() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;         // auto flush
-            int tail = 16;               // flushed on close
-            int arrayLen = (1 << 6) + 6; // increase this for stress tests
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 TINYINT[" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    byte[] arr = new byte[arrayLen];
-                    for (byte j = 0; j < arrayLen; j++) {
-                        arr[j] = (byte) ((i % 8) + j);
-                    }
-                    appender.beginRow().append(i).appendByteArray(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs =
-                     stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getInt(2), (i % 8) + arrayLen - 2);
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery("SELECT unnest(col2) FROM tab1 WHERE col1 = " + row)) {
-                for (byte j = 0; j < arrayLen; j++) {
-                    assertTrue(rs2.next());
-                    assertEquals(rs2.getByte(1), (byte) ((row % 8) + j));
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_nested_array_tinyint() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;         // auto flush
-            int tail = 17;               // flushed on close
-            int arrayLen = (1 << 6) + 5; // increase this for stress tests
-            int childLen = (1 << 6) + 7;
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 TINYINT[" + childLen + "][" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    byte[][] arr = new byte[arrayLen][childLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        for (int k = 0; k < childLen; k++) {
-                            arr[j][k] = (byte) (i + j + k);
-                        }
-                    }
-                    appender.beginRow().append(i).appendByteArray(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "][" + (childLen - 1) +
-                                                  "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getByte(2), (byte) (i + arrayLen + childLen - 4));
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery(
-                     "SELECT unnest(ucol2) FROM (SELECT unnest(col2) as ucol2 FROM tab1 WHERE col1 = " + row + ")")) {
-                for (int j = 0; j < arrayLen; j++) {
-                    for (int k = 0; k < childLen; k++) {
-                        assertTrue(rs2.next());
-                        assertEquals(rs2.getByte(1), (byte) (row + j + k));
-                    }
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_array_smallint() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;          // auto flush
-            int tail = 16;                // flushed on close
-            int arrayLen = (1 << 12) + 6; // increase this for stress tests
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 SMALLINT[" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    short[] arr = new short[arrayLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        arr[j] = (short) (i + j);
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs =
-                     stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getShort(2), (short) (i + arrayLen - 2));
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery("SELECT unnest(col2) FROM tab1 WHERE col1 = " + row)) {
-                for (int j = 0; j < arrayLen; j++) {
-                    assertTrue(rs2.next());
-                    assertEquals(rs2.getShort(1), (short) (row + j));
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_nested_array_smallint() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;         // auto flush
-            int tail = 17;               // flushed on close
-            int arrayLen = (1 << 6) + 5; // increase this for stress tests
-            int childLen = (1 << 6) + 7;
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 SMALLINT[" + childLen + "][" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    short[][] arr = new short[arrayLen][childLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        for (int k = 0; k < childLen; k++) {
-                            arr[j][k] = (short) (i + j + k);
-                        }
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "][" + (childLen - 1) +
-                                                  "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getShort(2), (short) (i + arrayLen + childLen - 4));
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery(
-                     "SELECT unnest(ucol2) FROM (SELECT unnest(col2) as ucol2 FROM tab1 WHERE col1 = " + row + ")")) {
-                for (int j = 0; j < arrayLen; j++) {
-                    for (int k = 0; k < childLen; k++) {
-                        assertTrue(rs2.next());
-                        assertEquals(rs2.getShort(1), (short) (row + j + k));
-                    }
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_array_integer() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;          // auto flush
-            int tail = 16;                // flushed on close
-            int arrayLen = (1 << 12) + 6; // increase this for stress tests
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 INTEGER[" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    int[] arr = new int[arrayLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        arr[j] = i + j;
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs =
-                     stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getInt(2), i + arrayLen - 2);
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery("SELECT unnest(col2) FROM tab1 WHERE col1 = " + row)) {
-                for (int j = 0; j < arrayLen; j++) {
-                    assertTrue(rs2.next());
-                    assertEquals(rs2.getInt(1), row + j);
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_nested_array_basic_integer() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 INTEGER[2][3])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                appender.beginRow().append(41).append(new int[][] {{42, 43}, {44, 45}, {46, 47}}).endRow();
-                appender.beginRow().append(48).append(new int[][] {{49, 50}, null, {53, 54}}).endRow();
-                appender.beginRow()
-                    .append(55)
-                    .append(new int[][] {{56, 57}, {58, 59}, {60, 61}},
-                            new boolean[][] {{false, true}, {false, false}, {true, false}})
-                    .endRow();
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col1, unnest(col2) FROM tab1 WHERE col1 = 41")) {
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 41);
-                Object[] array1 = (Object[]) rs.getArray(2).getArray();
-                assertEquals(array1.length, 2);
-                assertEquals(array1[0], 42);
-                assertEquals(array1[1], 43);
-
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 41);
-                Object[] array2 = (Object[]) rs.getArray(2).getArray();
-                assertEquals(array2.length, 2);
-                assertEquals(array2[0], 44);
-                assertEquals(array2[1], 45);
-
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 41);
-                Object[] array3 = (Object[]) rs.getArray(2).getArray();
-                assertEquals(array3.length, 2);
-                assertEquals(array3[0], 46);
-                assertEquals(array3[1], 47);
-
-                assertFalse(rs.next());
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col1, unnest(col2) FROM tab1 WHERE col1 = 48")) {
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 48);
-                Object[] array1 = (Object[]) rs.getArray(2).getArray();
-                assertEquals(array1.length, 2);
-                assertEquals(array1[0], 49);
-                assertEquals(array1[1], 50);
-
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 48);
-                assertNull(rs.getObject(2));
-
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 48);
-                Object[] array3 = (Object[]) rs.getArray(2).getArray();
-                assertEquals(array3.length, 2);
-                assertEquals(array3[0], 53);
-                assertEquals(array3[1], 54);
-
-                assertFalse(rs.next());
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col1, unnest(col2) FROM tab1 WHERE col1 = 55")) {
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 55);
-                Object[] array1 = (Object[]) rs.getArray(2).getArray();
-                assertEquals(array1.length, 2);
-                assertEquals(array1[0], 56);
-                assertNull(array1[1]);
-
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 55);
-                Object[] array2 = (Object[]) rs.getArray(2).getArray();
-                assertEquals(array2.length, 2);
-                assertEquals(array2[0], 58);
-                assertEquals(array2[1], 59);
-
-                assertTrue(rs.next());
-                assertEquals(rs.getInt(1), 55);
-                Object[] array3 = (Object[]) rs.getArray(2).getArray();
-                assertEquals(array3.length, 2);
-                assertNull(array3[0]);
-                assertEquals(array3[1], 61);
-
-                assertFalse(rs.next());
-            }
-        }
-    }
-
-    public static void test_appender_nested_array_integer() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 12;         // auto flush twice
-            int tail = 17;               // flushed on close
-            int arrayLen = (1 << 6) + 5; // increase this for stress tests
-            int childLen = (1 << 6) + 7;
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 INTEGER[" + childLen + "][" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    int[][] arr = new int[arrayLen][childLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        for (int k = 0; k < childLen; k++) {
-                            arr[j][k] = (i + 1) * (j + 1) * (k + 1);
-                        }
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "][" + (childLen - 1) +
-                                                  "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getInt(2), (i + 1) * (arrayLen - 1) * (childLen - 1));
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery(
-                     "SELECT unnest(ucol2) FROM (SELECT unnest(col2) as ucol2 FROM tab1 WHERE col1 = " + row + ")")) {
-                for (int j = 0; j < arrayLen; j++) {
-                    for (int k = 0; k < childLen; k++) {
-                        assertTrue(rs2.next());
-                        assertEquals(rs2.getInt(1), (row + 1) * (j + 1) * (k + 1));
-                    }
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_array_bigint() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;          // auto flush
-            int tail = 16;                // flushed on close
-            int arrayLen = (1 << 12) + 6; // increase this for stress tests
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 BIGINT[" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    long[] arr = new long[arrayLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        arr[j] = i + j;
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs =
-                     stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getLong(2), (long) (i + arrayLen - 2));
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery("SELECT unnest(col2) FROM tab1 WHERE col1 = " + row)) {
-                for (int j = 0; j < arrayLen; j++) {
-                    assertTrue(rs2.next());
-                    assertEquals(rs2.getLong(1), (long) (row + j));
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_nested_array_bigint() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;         // auto flush
-            int tail = 17;               // flushed on close
-            int arrayLen = (1 << 6) + 5; // increase this for stress tests
-            int childLen = (1 << 6) + 7;
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 BIGINT[" + childLen + "][" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    long[][] arr = new long[arrayLen][childLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        for (int k = 0; k < childLen; k++) {
-                            arr[j][k] = (long) (i + 1) * (j + 1) * (k + 1);
-                        }
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "][" + (childLen - 1) +
-                                                  "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getLong(2), (long) ((i + 1) * (arrayLen - 1) * (childLen - 1)));
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery(
-                     "SELECT unnest(ucol2) FROM (SELECT unnest(col2) as ucol2 FROM tab1 WHERE col1 = " + row + ")")) {
-                for (int j = 0; j < arrayLen; j++) {
-                    for (int k = 0; k < childLen; k++) {
-                        assertTrue(rs2.next());
-                        assertEquals(rs2.getLong(1), (long) (row + 1) * (j + 1) * (k + 1));
-                    }
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_array_float() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;          // auto flush
-            int tail = 16;                // flushed on close
-            int arrayLen = (1 << 12) + 6; // increase this for stress tests
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 FLOAT[" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    float[] arr = new float[arrayLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        arr[j] = (float) (i + j + 0.001);
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs =
-                     stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getFloat(2), (float) (i + arrayLen - 2 + 0.001));
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery("SELECT unnest(col2) FROM tab1 WHERE col1 = " + row)) {
-                for (int j = 0; j < arrayLen; j++) {
-                    assertTrue(rs2.next());
-                    assertEquals(rs2.getFloat(1), (float) (row + j + 0.001));
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_nested_array_float() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 12;         // auto flush twice
-            int tail = 17;               // flushed on close
-            int arrayLen = (1 << 6) + 5; // increase this for stress tests
-            int childLen = (1 << 6) + 7;
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 FLOAT[" + childLen + "][" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    float[][] arr = new float[arrayLen][childLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        for (int k = 0; k < childLen; k++) {
-                            arr[j][k] = (i + 1) * (j + 1) * (k + 1) + 0.001f;
-                        }
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "][" + (childLen - 1) +
-                                                  "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getFloat(2), (i + 1) * (arrayLen - 1) * (childLen - 1) + 0.001f);
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery(
-                     "SELECT unnest(ucol2) FROM (SELECT unnest(col2) as ucol2 FROM tab1 WHERE col1 = " + row + ")")) {
-                for (int j = 0; j < arrayLen; j++) {
-                    for (int k = 0; k < childLen; k++) {
-                        assertTrue(rs2.next());
-                        assertEquals(rs2.getFloat(1), (row + 1) * (j + 1) * (k + 1) + 0.001f);
-                    }
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_array_double() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;          // auto flush
-            int tail = 16;                // flushed on close
-            int arrayLen = (1 << 12) + 6; // increase this for stress tests
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 DOUBLE[" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    double[] arr = new double[arrayLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        arr[j] = (i + j + 0.001);
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs =
-                     stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getDouble(2), (i + arrayLen - 2 + 0.001));
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery("SELECT unnest(col2) FROM tab1 WHERE col1 = " + row)) {
-                for (int j = 0; j < arrayLen; j++) {
-                    assertTrue(rs2.next());
-                    assertEquals(rs2.getDouble(1), (row + j + 0.001));
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
-    public static void test_appender_nested_array_double() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 11;         // auto flush
-            int tail = 17;               // flushed on close
-            int arrayLen = (1 << 6) + 5; // increase this for stress tests
-            int childLen = (1 << 6) + 7;
-
-            stmt.execute("CREATE TABLE tab1(col1 INT, col2 DOUBLE[" + childLen + "][" + arrayLen + "])");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    double[][] arr = new double[arrayLen][childLen];
-                    for (int j = 0; j < arrayLen; j++) {
-                        for (int k = 0; k < childLen; k++) {
-                            arr[j][k] = (i + 1) * (j + 1) * (k + 1) + 0.001;
-                        }
-                    }
-                    appender.beginRow().append(i).append(arr).endRow();
-                }
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT col1, col2[" + (arrayLen - 1) + "][" + (childLen - 1) +
-                                                  "] FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    assertEquals(rs.getDouble(2), (i + 1) * (arrayLen - 1) * (childLen - 1) + 0.001);
-                }
-                assertFalse(rs.next());
-            }
-
-            int row = count - 2;
-            try (Statement stmt2 = conn.createStatement();
-                 ResultSet rs2 = stmt2.executeQuery(
-                     "SELECT unnest(ucol2) FROM (SELECT unnest(col2) as ucol2 FROM tab1 WHERE col1 = " + row + ")")) {
-                for (int j = 0; j < arrayLen; j++) {
-                    for (int k = 0; k < childLen; k++) {
-                        assertTrue(rs2.next());
-                        assertEquals(rs2.getDouble(1), (row + 1) * (j + 1) * (k + 1) + 0.001);
-                    }
-                }
-                assertFalse(rs2.next());
-            }
-        }
-    }
-
     public static void test_appender_roundtrip_blob() throws Exception {
         try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
              Statement stmt = conn.createStatement()) {
@@ -1591,231 +826,6 @@ public class TestAppender {
         }
     }
 
-    public static void test_appender_struct_basic() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-            stmt.execute("CREATE TABLE tab1 (col1 INTEGER, col2 STRUCT(s1 INTEGER, s2 VARCHAR))");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                long count = appender.beginRow()
-                                 .append(42)
-                                 .beginStruct()
-                                 .append(43)
-                                 .append("foo")
-                                 .endStruct()
-                                 .endRow()
-
-                                 .beginRow()
-                                 .append(44)
-                                 .beginStruct()
-                                 .append(45)
-                                 .appendNull()
-                                 .endStruct()
-                                 .endRow()
-
-                                 .beginRow()
-                                 .append(46)
-                                 .appendNull()
-                                 .endRow()
-                                 .flush();
-                assertEquals(count, 3L);
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1 WHERE col1 = 42")) {
-                assertTrue(rs.next());
-
-                assertEquals(rs.getInt(1), 42);
-                DuckDBStruct struct = (DuckDBStruct) rs.getObject(2);
-                Map<String, Object> map = struct.getMap();
-                assertEquals(map.get("s1"), 43);
-                assertEquals(map.get("s2"), "foo");
-
-                assertFalse(rs.next());
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1 WHERE col1 = 44")) {
-                assertTrue(rs.next());
-
-                assertEquals(rs.getInt(1), 44);
-                DuckDBStruct struct = (DuckDBStruct) rs.getObject(2);
-                Map<String, Object> map = struct.getMap();
-                assertEquals(map.get("s1"), 45);
-                assertNull(map.get("s2"));
-
-                assertFalse(rs.next());
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1 WHERE col1 = 46")) {
-                assertTrue(rs.next());
-
-                assertEquals(rs.getInt(1), 46);
-                assertNull(rs.getObject(2));
-
-                assertFalse(rs.next());
-            }
-        }
-    }
-
-    public static void test_appender_struct_nested() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-            stmt.execute(
-                "CREATE TABLE tab1 (col1 INTEGER, col2 STRUCT(s1 INTEGER, s2 STRUCT(ns1 INTEGER, ns2 VARCHAR)), col3 VARCHAR)");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                appender.beginRow()
-                    .append(42)
-                    .beginStruct()
-                    .append(43)
-                    .beginStruct()
-                    .append(44)
-                    .append("foo")
-                    .endStruct()
-                    .endStruct()
-                    .append("bar")
-                    .endRow()
-                    .flush();
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1 WHERE col1 = 42")) {
-                assertTrue(rs.next());
-
-                assertEquals(rs.getInt(1), 42);
-                DuckDBStruct struct = (DuckDBStruct) rs.getObject(2);
-                Map<String, Object> map = struct.getMap();
-                assertEquals(map.get("s1"), 43);
-                DuckDBStruct nested = (DuckDBStruct) map.get("s2");
-                Map<String, Object> nestedMap = nested.getMap();
-                assertEquals(nestedMap.get("ns1"), 44);
-                assertEquals(nestedMap.get("ns2"), "foo");
-                assertEquals(rs.getString(3), "bar");
-
-                assertFalse(rs.next());
-            }
-        }
-    }
-
-    public static void test_appender_struct_with_array() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-            stmt.execute("CREATE TABLE tab1 (col1 INTEGER, col2 STRUCT(s1 INTEGER, s2 INTEGER[2]))");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                long count = appender.beginRow()
-                                 .append(42)
-                                 .beginStruct()
-                                 .append(43)
-                                 .append(new int[] {44, 45})
-                                 .endStruct()
-                                 .endRow()
-                                 .flush();
-                assertEquals(count, 1L);
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1")) {
-                assertTrue(rs.next());
-
-                assertEquals(rs.getInt(1), 42);
-                DuckDBStruct struct = (DuckDBStruct) rs.getObject(2);
-                Map<String, Object> map = struct.getMap();
-                assertEquals(map.get("s1"), 43);
-                DuckDBArray arrayWrapper = (DuckDBArray) map.get("s2");
-                Object[] array = (Object[]) arrayWrapper.getArray();
-                assertEquals(array.length, 2);
-                assertEquals(array[0], 44);
-                assertEquals(array[1], 45);
-
-                assertFalse(rs.next());
-            }
-        }
-    }
-
-    public static void test_appender_struct_flush() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 12; // auto flush twice
-            int tail = 17;       // flushed on close
-
-            stmt.execute("CREATE TABLE tab1 (col1 INTEGER, col2 STRUCT(s1 INTEGER, s2 INTEGER[2], s3 VARCHAR))");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    appender.beginRow().append(i);
-                    appender.beginStruct().append(i + 1);
-                    if (0 == i % 7) {
-                        appender.append(new int[] {i + 2, i + 3}, new boolean[] {false, true});
-                    } else {
-                        appender.append(new int[] {i + 2, i + 3});
-                    }
-                    if (0 == i % 13) {
-                        appender.appendNull();
-                    } else {
-                        appender.append("foo" + i);
-                    }
-                    appender.endStruct();
-                    appender.endRow();
-                }
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    DuckDBStruct struct = (DuckDBStruct) rs.getObject(2);
-                    Map<String, Object> map = struct.getMap();
-                    assertEquals(map.get("s1"), i + 1);
-                    DuckDBArray arrayWrapper = (DuckDBArray) map.get("s2");
-                    Object[] array = (Object[]) arrayWrapper.getArray();
-                    assertEquals(array.length, 2);
-                    assertEquals(array[0], i + 2);
-                    if (0 == i % 7) {
-                        assertNull(array[1]);
-                    } else {
-                        assertEquals(array[1], i + 3);
-                    }
-                    if (0 == i % 13) {
-                        assertNull(map.get("s3"));
-                    } else {
-                        assertEquals(map.get("s3"), "foo" + i);
-                    }
-                }
-
-                assertFalse(rs.next());
-            }
-        }
-    }
-
-    public static void test_appender_struct_incomplete() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-            stmt.execute("CREATE TABLE tab1 (col1 INTEGER, col2 STRUCT(s1 INTEGER, s2 VARCHAR))");
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                // underflow
-                assertThrows(() -> {
-                    appender.beginRow()
-                        .append(42)
-                        .beginStruct()
-                        .append(43)
-                        //                            .append("foo")
-                        .endStruct();
-                }, SQLException.class);
-            }
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                // overflow
-                assertThrows(() -> {
-                    appender.beginRow()
-                        .append(42)
-                        .beginStruct()
-                        .append(43)
-                        .append("foo")
-                        .append(44) // extra field
-                        .endStruct();
-                }, SQLException.class);
-            }
-        }
-    }
-
     public static void test_appender_begin_misuse() throws Exception {
         try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
              Statement stmt = conn.createStatement()) {
@@ -1835,209 +845,6 @@ public class TestAppender {
             stmt.execute("CREATE TABLE tab1 (col1 INTEGER, col2 STRUCT(s1 INTEGER, s2 VARCHAR))");
             try (DuckDBAppender appender = conn.createAppender("tab1")) {
                 assertThrows(() -> { appender.beginRow().append(42).flush(); }, SQLException.class);
-            }
-        }
-    }
-
-    public static void test_appender_union_basic() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-            stmt.execute("CREATE TABLE tab1 (col1 INTEGER, col2 UNION(u1 INTEGER, u2 VARCHAR))");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                long count = appender.beginRow()
-                                 .append(42)
-                                 .beginUnion("u1")
-                                 .append(43)
-                                 .endUnion()
-                                 .endRow()
-
-                                 .beginRow()
-                                 .append(44)
-                                 .beginUnion("u1")
-                                 .appendNull()
-                                 .endUnion()
-                                 .endRow()
-
-                                 .beginRow()
-                                 .append(45)
-                                 .beginUnion("u2")
-                                 .append("foo")
-                                 .endUnion()
-                                 .endRow()
-
-                                 .flush();
-                assertEquals(count, 3L);
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1 WHERE col1 = 42")) {
-                assertTrue(rs.next());
-
-                assertEquals(rs.getInt(1), 42);
-                Object obj = rs.getObject(2);
-                assertTrue(obj instanceof Integer);
-                assertEquals(obj, 43);
-
-                assertFalse(rs.next());
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1 WHERE col1 = 44")) {
-                assertTrue(rs.next());
-
-                assertEquals(rs.getInt(1), 44);
-                Object obj = rs.getObject(2);
-                assertNull(obj);
-                //                assertTrue(rs.wasNull());
-
-                assertFalse(rs.next());
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1 WHERE col1 = 45")) {
-                assertTrue(rs.next());
-
-                assertEquals(rs.getInt(1), 45);
-                Object obj = rs.getObject(2);
-                assertTrue(obj instanceof String);
-                assertEquals(obj, "foo");
-
-                assertFalse(rs.next());
-            }
-        }
-    }
-
-    public static void test_appender_union_flush() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-
-            int count = 1 << 12; // auto flush twice
-            int tail = 17;       // flushed on close
-
-            stmt.execute("CREATE TABLE tab1 (col1 INTEGER, col2 UNION(u1 INTEGER, u2 INTEGER[2], u3 VARCHAR))");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    appender.beginRow().append(i);
-                    switch (i % 3) {
-                    case 0:
-                        appender.beginUnion("u1").append(i + 1);
-                        break;
-                    case 1:
-                        appender.beginUnion("u2").append(new int[] {i + 2, i + 3});
-                        break;
-                    default:
-                        appender.beginUnion("u3").append("foo" + i);
-                    }
-                    appender.endUnion();
-                    appender.endRow();
-                }
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1 ORDER BY col1")) {
-                for (int i = 0; i < count + tail; i++) {
-                    assertTrue(rs.next());
-                    assertEquals(rs.getInt(1), i);
-                    Object obj = rs.getObject(2);
-                    switch (i % 3) {
-                    case 0:
-                        assertTrue(obj instanceof Integer);
-                        assertEquals(obj, i + 1);
-                        break;
-                    case 1:
-                        assertTrue(obj instanceof DuckDBArray);
-                        DuckDBArray arrayWrapper = (DuckDBArray) obj;
-                        Object[] array = (Object[]) arrayWrapper.getArray();
-                        assertEquals(array.length, 2);
-                        assertEquals(array[0], i + 2);
-                        assertEquals(array[1], i + 3);
-                        break;
-                    default:
-                        assertTrue(obj instanceof String);
-                        assertEquals(obj, "foo" + i);
-                    }
-                }
-
-                assertFalse(rs.next());
-            }
-        }
-    }
-
-    public static void test_appender_union_nested() throws Exception {
-        try (DuckDBConnection conn = DriverManager.getConnection(JDBC_URL).unwrap(DuckDBConnection.class);
-             Statement stmt = conn.createStatement()) {
-            stmt.execute("CREATE TABLE tab1 ("
-                         + "col1 INTEGER, "
-                         + "col2 STRUCT( "
-                         + "   s1 INTEGER, "
-                         + "   s2 UNION("
-                         + "       u1 INTEGER,"
-                         + "       u2 STRUCT("
-                         + "           us1 INTEGER, "
-                         + "           us2 INTEGER[2],"
-                         + "           us3 VARCHAR"
-                         + "       )"
-                         + "   )"
-                         + "), "
-                         + "col3 VARCHAR)");
-
-            try (DuckDBAppender appender = conn.createAppender("tab1")) {
-                appender.beginRow()
-                    .append(42)
-                    .beginStruct()
-                    .append(43)
-                    .beginUnion("u1")
-                    .append(44)
-                    .endUnion()
-                    .endStruct()
-                    .append("foo")
-                    .endRow()
-
-                    .beginRow()
-                    .append(45)
-                    .beginStruct()
-                    .append(46)
-                    .beginUnion("u2")
-                    .beginStruct()
-                    .append(47)
-                    .append(new int[] {48, 49})
-                    .append("bar")
-                    .endStruct()
-                    .endUnion()
-                    .endStruct()
-                    .append("baz")
-                    .endRow();
-            }
-
-            try (ResultSet rs = stmt.executeQuery("SELECT * FROM tab1 ORDER BY col1")) {
-                assertTrue(rs.next());
-
-                assertEquals(rs.getInt(1), 42);
-                DuckDBStruct struct = (DuckDBStruct) rs.getObject(2);
-                Map<String, Object> map = struct.getMap();
-                assertEquals(map.size(), 2);
-                assertEquals(map.get("s1"), 43);
-                assertEquals(map.get("s2"), 44);
-                assertEquals(rs.getString(3), "foo");
-
-                assertTrue(rs.next());
-
-                assertEquals(rs.getInt(1), 45);
-                DuckDBStruct struct1 = (DuckDBStruct) rs.getObject(2);
-                Map<String, Object> map1 = struct1.getMap();
-                assertEquals(map1.size(), 2);
-                assertEquals(map1.get("s1"), 46);
-                DuckDBStruct struct2 = (DuckDBStruct) map1.get("s2");
-                Map<String, Object> map2 = struct2.getMap();
-                assertEquals(map2.size(), 3);
-                assertEquals(map2.get("us1"), 47);
-                DuckDBArray arrWrapper = (DuckDBArray) map2.get("us2");
-                Object[] arr = (Object[]) arrWrapper.getArray();
-                assertEquals(arr.length, 2);
-                assertEquals(arr[0], 48);
-                assertEquals(arr[1], 49);
-                assertEquals(map2.get("us3"), "bar");
-                assertEquals(rs.getString(3), "baz");
-
-                assertFalse(rs.next());
             }
         }
     }
